move connect to compute
